Rename factory as App
