+++ src/cpp/exe/outer_loop/main.cpp
@@ -1,5 +1,5 @@
 #include "antares-xpansion/benders/benders_mpi/common_mpi.h"
-#include "antares-xpansion/benders/factories/BendersFactory.h"
+#include "antares-xpansion/benders/factories/BendersApp.h"
 
 int main(int argc, char** argv)
 {
